requirements.txt
sqlalchemy
pypdf
pdfminer.six
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
schedule
python-dotenv
pyyaml
